Vello Hybrid: Move `RenderContext` into `examples/common` (#860)
Exposing `RenderContext` publicly from Vello is one of the greatest
mistakes, as it's a significant layering break. However, it does
simplify the multi-GPU code we have for our examples significantly, so
removing it entirely is quite an involved process. For now, we can move
it to `examples/common` in Vello Hybrid.

(I tend to think that our "simple" example is an anti-pattern; it's not
our job to teach people how to set up wgpu, and getting that entire
integration exactly right is *really hard*; to call an example "simple"
is the same as calling it "wrong". There absolutely is a role for
"idiomatic wgpu + winit integration", but trying to fit it in a single
file whilst also teaching our API is not really conducive to that)
