+++ sparse_strips/vello_hybrid/src/util.rs
@@ -7,191 +7,6 @@
 
 use std::ops::RangeInclusive;
 
-use wgpu::{
-    Adapter, Device, Features, Instance, Limits, MemoryHints, Queue, Surface, SurfaceConfiguration,
-    SurfaceTarget, TextureFormat,
-};
-
-/// Simple render context that maintains wgpu state for rendering the pipeline.
-#[derive(Debug)]
-pub struct RenderContext {
-    /// The instance of the wgpu instance
-    pub instance: Instance,
-    /// The devices of the wgpu instance
-    pub devices: Vec<DeviceHandle>,
-}
-
-/// A handle to a device
-#[derive(Debug)]
-pub struct DeviceHandle {
-    /// The adapter of the device
-    pub adapter: Adapter,
-    /// The device
-    pub device: Device,
-    /// The queue of the device
-    pub queue: Queue,
-}
-
-impl RenderContext {
-    /// Creates a new render context
-    #[expect(
-        clippy::new_without_default,
-        reason = "Creating a wgpu Instance is something which should only be done rarely"
-    )]
-    pub fn new() -> Self {
-        let backends = wgpu::Backends::from_env().unwrap_or_default();
-        let flags = wgpu::InstanceFlags::from_build_config().with_env();
-        let backend_options = wgpu::BackendOptions::from_env_or_default();
-        let instance = Instance::new(&wgpu::InstanceDescriptor {
-            backends,
-            flags,
-            backend_options,
-        });
-        Self {
-            instance,
-            devices: Vec::new(),
-        }
-    }
-
-    /// Creates a new surface for the specified window and dimensions.
-    pub async fn create_surface<'w>(
-        &mut self,
-        window: impl Into<SurfaceTarget<'w>>,
-        width: u32,
-        height: u32,
-        present_mode: wgpu::PresentMode,
-        format: TextureFormat,
-    ) -> RenderSurface<'w> {
-        self.create_render_surface(
-            self.instance
-                .create_surface(window.into())
-                .expect("Error creating surface"),
-            width,
-            height,
-            present_mode,
-            format,
-        )
-        .await
-    }
-
-    /// Creates a new render surface for the specified window and dimensions.
-    pub async fn create_render_surface<'w>(
-        &mut self,
-        surface: Surface<'w>,
-        width: u32,
-        height: u32,
-        present_mode: wgpu::PresentMode,
-        format: TextureFormat,
-    ) -> RenderSurface<'w> {
-        let dev_id = self
-            .device(Some(&surface))
-            .await
-            .expect("No compatible device");
-
-        let config = SurfaceConfiguration {
-            usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
-            format,
-            width,
-            height,
-            present_mode,
-            desired_maximum_frame_latency: 2,
-            alpha_mode: wgpu::CompositeAlphaMode::Auto,
-            view_formats: vec![],
-        };
-        let surface = RenderSurface {
-            surface,
-            config,
-            dev_id,
-        };
-        self.configure_surface(&surface);
-        surface
-    }
-
-    /// Resizes the surface to the new dimensions.
-    pub fn resize_surface(&self, surface: &mut RenderSurface<'_>, width: u32, height: u32) {
-        surface.config.width = width;
-        surface.config.height = height;
-        self.configure_surface(surface);
-    }
-
-    /// Sets the present mode for the surface
-    pub fn set_present_mode(
-        &self,
-        surface: &mut RenderSurface<'_>,
-        present_mode: wgpu::PresentMode,
-    ) {
-        surface.config.present_mode = present_mode;
-        self.configure_surface(surface);
-    }
-
-    fn configure_surface(&self, surface: &RenderSurface<'_>) {
-        let device = &self.devices[surface.dev_id].device;
-        surface.surface.configure(device, &surface.config);
-    }
-
-    /// Finds or creates a compatible device handle id.
-    pub async fn device(&mut self, compatible_surface: Option<&Surface<'_>>) -> Option<usize> {
-        let compatible = match compatible_surface {
-            Some(s) => self
-                .devices
-                .iter()
-                .enumerate()
-                .find(|(_, d)| d.adapter.is_surface_supported(s))
-                .map(|(i, _)| i),
-            None => (!self.devices.is_empty()).then_some(0),
-        };
-        if compatible.is_none() {
-            return self.new_device(compatible_surface).await;
-        }
-        compatible
-    }
-
-    /// Creates a compatible device handle id.
-    async fn new_device(&mut self, compatible_surface: Option<&Surface<'_>>) -> Option<usize> {
-        let adapter =
-            wgpu::util::initialize_adapter_from_env_or_default(&self.instance, compatible_surface)
-                .await?;
-        let limits = Limits::default();
-
-        let (device, queue) = adapter
-            .request_device(
-                &wgpu::DeviceDescriptor {
-                    label: None,
-                    required_features: Features::empty(),
-                    required_limits: limits,
-                    memory_hints: MemoryHints::default(),
-                },
-                None,
-            )
-            .await
-            .ok()?;
-        let device_handle = DeviceHandle {
-            adapter,
-            device,
-            queue,
-        };
-        self.devices.push(device_handle);
-        Some(self.devices.len() - 1)
-    }
-}
-
-/// Combination of surface and its configuration.
-#[derive(Debug)]
-pub struct RenderSurface<'s> {
-    /// The surface
-    pub surface: Surface<'s>,
-    /// The configuration of the surface
-    pub config: SurfaceConfiguration,
-    /// The device id
-    pub dev_id: usize,
-}
-
-struct NullWake;
-
-impl std::task::Wake for NullWake {
-    fn wake(self: std::sync::Arc<Self>) {}
-}
-
 /// Represents dimension constraints for surfaces
 #[derive(Debug)]
 pub struct DimensionConstraints {
